add security on the media admin controllers

+++ src/Bridge/EasyAdmin/src/Controller/MediaAdminController.php
@@ -13,6 +13,7 @@ use JoliCode\MediaBundle\Bridge\EasyAdmin\Form\Type\MoveType;
 use JoliCode\MediaBundle\Bridge\EasyAdmin\Form\Type\RenameDirectoryType;
 use JoliCode\MediaBundle\Bridge\EasyAdmin\Form\Type\RenameType;
 use JoliCode\MediaBundle\Bridge\EasyAdmin\Form\Type\UploadType;
+use JoliCode\MediaBundle\Bridge\Security\Voter\AdminAction;
 use JoliCode\MediaBundle\Conversion\Converter;
 use JoliCode\MediaBundle\Exception\ForbiddenPathException;
 use JoliCode\MediaBundle\Exception\MediaInUseException;
@@ -35,6 +36,7 @@ use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
+use Symfony\Component\Security\Core\Authorization\AuthorizationCheckerInterface;
 use Symfony\Contracts\Translation\TranslatorInterface;
 use Twig\Environment;
 
@@ -50,12 +52,17 @@ class MediaAdminController extends AbstractController
         private readonly Environment $twig,
         private readonly FormFactoryInterface $formFactory,
         private readonly AdminUrlGenerator $adminUrlGenerator,
+        private readonly ?AuthorizationCheckerInterface $authorizationChecker = null,
     ) {
     }
 
     #[Route(path: '/create-directory', name: 'create_directory', methods: [Request::METHOD_POST])]
     public function createDirectory(Request $request): RedirectResponse
     {
+        $this->denyAccessUnlessGranted(AdminAction::CREATE_DIRECTORY, new AdminAction(
+            libraryName: $this->getLibrary()->getName(),
+        ));
+
         $directory = null;
         $form = $this->createCreateDirectoryForm(null, null);
         $form->handleRequest($request);
@@ -67,6 +74,12 @@ class MediaAdminController extends AbstractController
         if ($form->isValid()) {
             $parentDirectory = $form->get('parentDirectory')->getData();
             $directory = $form->get('directory')->getData();
+
+            $this->denyAccessUnlessGranted(AdminAction::CREATE_DIRECTORY, new AdminAction(
+                libraryName: $this->getLibrary()->getName(),
+                path: $parentDirectory,
+                to: $directory,
+            ));
 
             if ('' !== $parentDirectory) {
                 $directory = \sprintf('%s/%s', $parentDirectory, $directory);
@@ -129,6 +142,11 @@ class MediaAdminController extends AbstractController
         if ($form->isSubmitted() && $form->isValid()) {
             $key = $form->get('path')->getData();
 
+            $this->denyAccessUnlessGranted(AdminAction::DELETE, new AdminAction(
+                libraryName: $this->getLibrary()->getName(),
+                path: Resolver::normalizePath($key),
+            ));
+
             // Remove the media and the cache files
             try {
                 $this->getOriginalStorage()->delete($key);
@@ -179,6 +197,11 @@ class MediaAdminController extends AbstractController
 
         if ($form->isSubmitted() && $form->isValid()) {
             $key = $form->get('path')->getData();
+
+            $this->denyAccessUnlessGranted(AdminAction::DELETE_DIRECTORY, new AdminAction(
+                libraryName: $this->getLibrary()->getName(),
+                path: Resolver::normalizePath($key),
+            ));
 
             try {
                 $this->getOriginalStorage()->deleteDirectory($key);
@@ -238,6 +261,10 @@ class MediaAdminController extends AbstractController
         }
 
         $currentKey = Resolver::normalizePath($key);
+        $this->denyAccessUnlessGranted(AdminAction::LIST, new AdminAction(
+            libraryName: $this->getLibrary()->getName(),
+            path: $currentKey,
+        ));
 
         try {
             $trashPath = $this->getOriginalStorage()->getTrashPath();
@@ -299,6 +326,12 @@ class MediaAdminController extends AbstractController
         $from = Resolver::normalizePath($form->get('from')->getData());
         $to = Resolver::normalizePath($form->get('to')->getData() ?? '');
 
+        $this->denyAccessUnlessGranted(AdminAction::MOVE, new AdminAction(
+            libraryName: $this->getLibrary()->getName(),
+            path: $from,
+            to: $to,
+        ));
+
         if ($form->isValid()) {
             try {
                 $target = \sprintf('%s/%s', $to, basename($from));
@@ -343,6 +376,12 @@ class MediaAdminController extends AbstractController
     #[Route(path: '/regenerate/{variation}/{key}', name: 'regenerate_variation', requirements: ['key' => '.+'], methods: [Request::METHOD_GET])]
     public function regenerateVariation(string $key, string $variation): RedirectResponse
     {
+        $this->denyAccessUnlessGranted(AdminAction::REGENERATE_VARIATION, new AdminAction(
+            libraryName: $this->getLibrary()->getName(),
+            path: Resolver::normalizePath($key),
+            variation: $variation,
+        ));
+
         $media = $this->resolver->resolveMedia($key);
         $variation = $this->getLibrary()->getVariation($variation);
         $mediaVariation = $variation->getForMedia($media);
@@ -377,6 +416,12 @@ class MediaAdminController extends AbstractController
 
         $from = Resolver::normalizePath((string) $form->get('from')->getData());
         $to = Resolver::normalizePath($form->get('to')->getData());
+
+        $this->denyAccessUnlessGranted(AdminAction::RENAME_DIRECTORY, new AdminAction(
+            libraryName: $this->getLibrary()->getName(),
+            path: $from,
+            to: $to,
+        ));
 
         if ($form->isValid()) {
             try {
@@ -428,6 +473,11 @@ class MediaAdminController extends AbstractController
     #[Route(path: '/show/{key}', name: 'show', requirements: ['key' => '.+'], methods: [Request::METHOD_GET, Request::METHOD_POST])]
     public function show(AdminContext $adminContext, Request $request, string $key): Response
     {
+        $this->denyAccessUnlessGranted(AdminAction::SHOW, new AdminAction(
+            libraryName: $this->getLibrary()->getName(),
+            path: Resolver::normalizePath($key),
+        ));
+
         $media = $this->resolver->resolveMedia($key);
         $key = $media->getPath();
         $this->addAssets($adminContext);
@@ -444,6 +494,11 @@ class MediaAdminController extends AbstractController
             if ($renameFileForm->isValid()) {
                 try {
                     $target = \sprintf('%s/%s', $media->getFolderPath(), $to);
+                    $this->denyAccessUnlessGranted(AdminAction::MOVE, new AdminAction(
+                        libraryName: $this->getLibrary()->getName(),
+                        path: Resolver::normalizePath($key),
+                        to: $target,
+                    ));
                     $this->getOriginalStorage()->move($key, $target);
                     $this->addFlash(
                         'success',
@@ -509,6 +564,10 @@ class MediaAdminController extends AbstractController
     #[Route(path: '/upload', name: 'upload', methods: [Request::METHOD_POST])]
     public function upload(Request $request): JsonResponse
     {
+        $this->denyAccessUnlessGranted(AdminAction::UPLOAD, new AdminAction(
+            libraryName: $this->getLibrary()->getName(),
+        ));
+
         $form = $this->createUploadForm();
         $form->handleRequest($request);
 
@@ -522,6 +581,11 @@ class MediaAdminController extends AbstractController
             $filename = $uploadedFile->getClientOriginalName();
             $key = \sprintf('%s/%s', $form->get('path')->getData(), $filename);
             $storage = $this->getOriginalStorage();
+
+            $this->denyAccessUnlessGranted(AdminAction::UPLOAD, new AdminAction(
+                libraryName: $this->getLibrary()->getName(),
+                path: Resolver::normalizePath($key),
+            ));
 
             if ($storage->has($key)) {
                 throw new UploadException($this->translator->trans('media.upload.error.already_exists', ['%file%' => $filename], 'JoliMediaEasyAdminBundle'));
@@ -580,6 +644,21 @@ class MediaAdminController extends AbstractController
         return new JsonResponse([
             'error' => $message,
         ], Response::HTTP_BAD_REQUEST);
+    }
+
+    protected function denyAccessUnlessGranted(mixed $attribute, mixed $subject = null, string $message = 'Access Denied.'): void
+    {
+        if (!$this->authorizationChecker instanceof AuthorizationCheckerInterface) {
+            return;
+        }
+
+        if (!$this->authorizationChecker->isGranted($attribute, $subject)) {
+            $e = $this->createAccessDeniedException($message);
+            $e->setAttributes([$attribute]);
+            $e->setSubject($subject);
+
+            throw $e;
+        }
     }
 
     private function addAssets(AdminContext $adminContext): void
